Module asyncio::ip

Structs

Icmp
The Internet Control Message Protocol (v6).

IpAddrV4
Implements IP version 4 style addresses.

IpAddrV6
Implements IP version 6 style addresses.

IpEndpoint
The endpoint of internet protocol.

LlAddr
Implements Link-layer addresses.

MulticastEnableLoopback
Socket option determining whether outgoing multicast packets will be received on the same socket if it is a member of the multicast group.

MulticastHops
Socket option for time-to-live associated with outgoing multicast packets.

MulticastJoinGroup
Socket option to join a multicast group on a specified interface.

MulticastLeaveGroup
Socket option to leave a multicast group on a specified interface.

NoDelay
Socket option for disabling the Nagle algorithm.

OutboundInterface
Socket option for local interface to use for outgoing multicast packets.

Passive
A query of the resolver for the passive mode.

Resolver
An entry produced by a resolver.

ResolverIter
An iterator over the entries produced by a resolver.

Tcp
The Transmission Control Protocol.

Udp
The User Datagram Protocol.

UnicastHops
Socket option for time-to-live associated with outgoing unicast packets.

V6Only
Socket option for get/set an IPv6 socket supports IPv6 communication only.

Enums

IpAddr
implements version-independent IP addresses.

Traits

IpProtocol
A category of an internet protocol.

ResolverQuery
A query to be passed to a resolver.

ToEndpoint
Provides conversion to a IP-endpoint.

Functions

host_name
Get the current host name.

Type Definitions

IcmpEndpoint
The ICMP(v6) endpoint type.

IcmpResolver
The ICMP(v6) resolver type.

IcmpSocket
The ICMP(v6) socket type.

TcpEndpoint
The TCP endpoint type.

TcpListener
The TCP listener type.

TcpResolver
The TCP resolver type.

TcpSocket
The TCP socket type.

UdpEndpoint
The UDP endpoint type.

UdpResolver
The UDP resolver type.

UdpSocket
The UDP socket type.