Module asyncio::socket_base

Structs

Broadcast
Socket option to permit sending of broadcast messages.

BytesReadable
IO control command to get the amount of data that can be read without blocking.

BytesWritten

Debug
Socket option to enable socket-level debugging.

DoNotRoute
Socket option to don't use a gateway. send to local network host only.

Ifreq
IO control command to interface request data.

IfreqSocket
IO control command socket for Ifreq.

KeepAlive
Socket option to send keep-alives.

Linger
Socket option to specify whether the socket lingers on close if unsent data is present.

RecvBufferSize
Socket option for the receive buffer size of a socket.

RecvLowWatermark
Socket option for the receive low watermark.

ReuseAddr
Socket option to allow the socket to be bound to an address that is already in use.

SendBufferSize
Socket option for the send buffer size of a socket.

SendLowWatermark
Socket option for the send low watermark.

Traits

IfreqName

Type Definitions

IfreqAddr

IfreqBroadcast

IfreqDestinate

IfreqGetFlags

IfreqGetHwAddr

IfreqGetIndex

IfreqGetMTUSize

IfreqNetmask

IfreqSetFlags

IfreqSetHwAddr

IfreqSetMTUSize