Struct asyncio::StreamBuf

pub struct StreamBuf { /* fields omitted */ }

Automatically resizing buffer.

Methods

impl StreamBuf

fn new() -> StreamBuf

Returns a new StreamBuf.

Equivalent to with_max_len(usize::max_len())

Examples

use asyncio::StreamBuf;

let mut sbuf = StreamBuf::new();

fn with_max_len(max: usize) -> StreamBuf

Returns a new StreamBuf with the max length of the allocatable size.

Examples

use asyncio::StreamBuf;

let mut sbuf = StreamBuf::with_max_len(1024);

fn capacity(&self) -> usize

Returns an allocated size of the buffer.

Examples

use asyncio::StreamBuf;

let mut sbuf = StreamBuf::new();
assert_eq!(sbuf.capacity(), 0);

fn clear(&mut self)

Clears the buffer, removing all values.

Examples

use asyncio::StreamBuf;

let mut sbuf = StreamBuf::from(vec![1,2,3]);
sbuf.clear();
assert_eq!(sbuf.is_empty(), true);

fn consume(&mut self, len: usize)

Remove characters from the input sequence.

Examples

use asyncio::StreamBuf;

let mut sbuf = StreamBuf::from(vec![1,2,3]);
assert_eq!(sbuf.len(), 3);
sbuf.consume(3);
assert_eq!(sbuf.len(), 0);

fn commit(&mut self, len: usize)

Move characters from the output sequence to the input sequence.

Examples

use asyncio::StreamBuf;

let mut sbuf = StreamBuf::new();
let _ = sbuf.prepare(256);
assert_eq!(sbuf.len(), 0);
sbuf.commit(3);
assert_eq!(sbuf.len(), 3);

fn is_empty(&self) -> bool

Returns true if the empty buffer.

Examples

use asyncio::StreamBuf;

let sbuf = StreamBuf::new();
assert!(sbuf.is_empty());

fn len(&self) -> usize

Returns a length of the input sequence.

Examples

use asyncio::StreamBuf;

let sbuf = StreamBuf::from(vec![1,2,3]);
assert_eq!(sbuf.len(), 3);

fn max_len(&self) -> usize

Returns a maximum length of the StreamBuf.

Examples

use asyncio::StreamBuf;

let sbuf = StreamBuf::new();
assert_eq!(sbuf.max_len(), usize::max_value());

fn prepare(&mut self, len: usize) -> Result<&mut [u8]>

Returns a &mut [u8] that represents a output sequence.

Examples

use asyncio::StreamBuf;

let mut sbuf = StreamBuf::with_max_len(8);
assert_eq!(sbuf.prepare(5).unwrap().len(), 5);
sbuf.commit(5);
assert_eq!(sbuf.prepare(5).unwrap().len(), 3);

fn prepare_exact(&mut self, len: usize) -> Result<&mut [u8]>

Returns a &mut [u8] that represents a output sequence.

Examples

use asyncio::StreamBuf;

let mut sbuf = StreamBuf::with_max_len(8);
assert_eq!(sbuf.prepare_exact(5).unwrap().len(), 5);
sbuf.commit(5);
assert!(sbuf.prepare_exact(5).is_err());

fn as_slice(&self) -> &[u8]

Returns a &[u8] that represents the input sequence.

fn as_mut_slice(&mut self) -> &mut [u8]

Returns a &mut [u8] that represents the input sequence.